requirements.txt
python-telegram-bot
openai
python-dotenv
nltk
pymongo
twilio
langchain